Empty shell of the SFView Avalonia port
It only has MainWindow.xaml and starts with a black window. The port is named StarwingMapVisualizer. Code files from the original application are copied and their namespace changed because they are modified (and this include changes in method signatures) in the porting process and I need a way to distinguish variants.